+++ R/rust_result.R
@@ -95,67 +95,7 @@ or_else = function(x, f) {
 }
 
 
-#' rust-like unwrapping of result. Useful to keep error handling on the R side.
-#'
-#' @param result a list here either element ok or err is NULL, or both if ok is litteral NULL
-#' @param call context of error or string
-#' @param context a msg to prefix a raised error with
-#'
-#' @return the ok-element of list , or a error will be thrown
-#' @keywords internal
-#' @export
-#'
-#' @examples
-#'
-#' structure(list(ok = "foo", err = NULL), class = "extendr_result")
-#'
-#' tryCatch(
-#'   unwrap(
-#'     structure(
-#'       list(ok = NULL, err = "something happen on the rust side"),
-#'       class = "extendr_result"
-#'     )
-#'   ),
-#'   error = function(err) as.character(err)
-#' )
-unwrap = function(result, context = NULL, call = sys.call(1L)) {
-  # if not a result
-  if (!is_result(result)) {
-    stopf("Internal error: cannot unwrap non result")
-  }
-
-  # if result is ok (ok can be be valid null, hence OK if both ok and err is null)
-  if (is.null(result$err)) {
-    return(result$ok)
-  }
-
-  # if result is error, make a pretty with context
-  if (is.null(result$ok) && !is.null(result$err)) {
-    if (!is.null(context)) {
-      result$err = paste(context, result$err)
-    }
-
-    stop(
-      paste(
-        result$err,
-        if (!polars_optenv$do_not_repeat_call) {
-          paste(
-            "\n when calling :\n",
-            paste(capture.output(print(call)), collapse = "\n")
-          )
-        }
-      ),
-      domain = NA,
-      call. = FALSE
-    )
-  }
-
-  # if not ok XOR error, then roll over
-  stopf("Internal error: result object corrupted")
-}
-
-
-#' Internal preferred function to throw errors
+#' pstop
 #' @description DEPRECATED USE stopf instead
 #' @param err error msg string
 #' @param call calling context
@@ -168,15 +108,4 @@ unwrap = function(result, context = NULL, call = sys.call(1L)) {
 #' tryCatch(f(), error = \(e) as.character(e))
 pstop = function(err, call = sys.call(1L)) {
   unwrap(list(ok = NULL, err = err), call = call)
-}
-
-# capture error in any R side arguments, and pass to rust side to preserve context and write
-# really sweet error messages
-result = function(x, msg = "an error because:\n") {
-  tryCatch(
-    Ok(x),
-    error = function(err) {
-      Err(paste0(msg, err$message))
-    }
-  )
 }